fixed various compiler warnings (gcc and clang)

+++ include/rbdl/Contacts.h
@@ -108,7 +108,7 @@ struct RBDL_DLLAPI ConstraintSet {
 	bool Bind (const Model &model);
 
 	/** \brief Returns the number of constraints. */
-	unsigned int size() const {
+	size_t size() const {
 		return acceleration.size();
 	}
